Upgrade Benchmarks and Tests
