+++ Benchmarks/Program.cs
@@ -1,5 +1,7 @@
 ﻿using BenchmarkDotNet.Running;
-using CombatlogParser;
+using CombatlogParser.Parsing;
+
+namespace Benchmarks;
 
 public class Program
 {
@@ -12,21 +14,21 @@ public class Program
 [SimpleJob]
 public class SubeventParse
 {
-	private string[] subevents = new string[]
-	{
+	private readonly string[] subevents =
+	[
 		"SPELL_DAMAGE",
 		"SPELL_AURA_APPLIED",
 		"SPELL_PERIODIC_DAMAGE_SUPPORT",
 		"RANGE_DAMAGE",
 		"SWING_DAMAGE_LANDED"
-	};
+	];
 
 	[Benchmark]
 	public void TryParse()
 	{
 		foreach (var subevent in subevents)
 		{
-			ParsingUtil.TryParsePrefixSuffixSubeventF(subevent, out _, out _);
+			_ = ParsingUtil.TryParsePrefixSuffixSubeventF(subevent, out _, out _);
 		}
 	}
 }
