+++ Tests/DBTests.cs
@@ -1,0 +1,6 @@
+﻿namespace Tests
+{
+	public class DBTests
+	{
+	}
+}
