--- Tests/Tests/DBTests.cs
@@ -1,6 +1,0 @@
-﻿namespace CombatlogParser.Tests
-{
-	public class DBTests
-	{
-	}
-}
